+++ volmdlr/faces.py
@@ -43,7 +43,8 @@ def octree_face_decomposition(face):
     Decomposes the face discretization triangle faces inside eight boxes from a bounding box octree structure.
 
     :param face: given face.
-    :return: returns a dictionary containing bounding boxes as keys and as values, a list of faces inside that bbox.
+    :return: returns a dictionary containing bounding boxes as keys and as values, a list of faces
+    inside that bounding box.
     """
     triangulation = face.triangulation()
     triangulation_faces = triangulation.faces