+++ CHANGELOG.md
@@ -19,11 +19,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - BSplineCurve3D: trim
 - FullArc3D: hash
 - SphericalSurface3D: enhance repair_periodicity_method
+- CylindricalSurface3D: concurrent_plane_intersection
+- BSplineFace3D: fix neutral_fiber
 - Step: assembly import
 - BSplineFace3D: fix neutral_fiber
 - BSplineSurface3D: improve bsplinecurve3d_to_2d.
+- BSplineSurface3D: improve bsplinecurve3d_to_3d.
+- Circle2D: plot
+- Line3D: fix Line3D plot()
+- Vector2D: plot()
+- fix RevolutionFace3D init parameter wire to edge.
+- fix Sweep: bug when first primitive is an arc.
 - Step.py: enhance step import/export
-
 
 ### Refactor
 - refator some classes' init in primitives3D. 
@@ -31,6 +38,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Composite_primitives
 - Surface3D: enhance repair_primitives_periodicity method.
 - BSplineCurve: replace periodic bool parameter with verification inside from_points_intepolation method.
+- Wire3D: removes heritage from volmdlr.core.CompositePrimitive3D
 
 ### Changed
 - Moves functions from step.py to volmdlr.utils.step_reader